services/transcription/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
torch>=2.0.0
python-docx>=0.8.11
numpy>=1.24.0
